refactor: criando pacote session em app + fix: correção de chamada incorreta no GameController
